requirements.txt
fastapi>=0.115.4,<0.115.6
uvicorn>=0.15.0,<0.16.0
python-dotenv
pytest
spacy